+++ src/classes/Bot.ts
@@ -1,19 +1,36 @@
+import env from 'env-var'
 import express from 'express'
-import Discord, { Client, Intents } from 'discord.js'
-import MediaManager from './MediaManager'
-import Api from '../api'
-import Config from './Config'
+import Discord, { Client, GatewayIntentBits, Partials } from 'discord.js'
+import MediaManager from './MediaManager.js'
+import Api from '../api.js'
+import Config from './Config.js'
+import Cache from '../types/Cache.js'
+import Throttler from './Throttler.js'
 
 export default class Bot {
   config: Config
   discord: Client
   mediaManager: MediaManager
-  api: express.Application
+  api?: express.Application
+  cache: Cache
 
   constructor(discord: Client, mediaManager: MediaManager, config: Config) {
     this.discord = discord
     this.mediaManager = mediaManager
     this.config = config
+    this.cache = this.getInitialCache()
+  }
+
+  // Initialize and return cache instance
+  protected getInitialCache(): Cache {
+    const { play, random } = this.config.data.app.commandThrottling ?? {}
+
+    return {
+      throttles: {
+        play: play ? new Throttler(play.usages, play.duration) : undefined,
+        random: random ? new Throttler(random.usages, random.duration) : undefined,
+      },
+    }
   }
 
   async getOwner() {
@@ -25,17 +42,19 @@ export default class Bot {
   }
 
   static async start(config: Config): Promise<Bot> {
-    const mediaManager = await MediaManager.init(process.env.MEDIA_FOLDER)
+    const mediaManager = await MediaManager.init(env.get('MEDIA_FOLDER').required().asString())
     const discord = new Discord.Client({
       intents: [
-        Intents.FLAGS.GUILDS,
-        Intents.FLAGS.GUILD_VOICE_STATES,
-        Intents.FLAGS.GUILD_MESSAGES,
-        Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
-        Intents.FLAGS.DIRECT_MESSAGES,
+        GatewayIntentBits.Guilds,
+        GatewayIntentBits.GuildVoiceStates,
+        GatewayIntentBits.GuildMessages,
+        GatewayIntentBits.GuildMessageReactions,
+        GatewayIntentBits.DirectMessages,
       ],
-      partials: ['CHANNEL'],
+      partials: [Partials.Channel],
     })
+
+    console.info(`${ mediaManager.medias.length } media(s) found`)
     await discord.login(process.env.DISCORD_TOKEN)
 
     const app = new Bot(discord, mediaManager, config)