+++ src/slashCommands/play.ts
@@ -1,33 +1,39 @@
+import { Client, StageChannel, VoiceChannel } from 'discord.js'
 import {
   AutocompleteChoice,
   AutocompleteContext,
   ButtonStyle,
+  CommandContext,
   CommandOptionType,
+  ComponentContext,
   ComponentType,
-  SlashCommand,
+  SlashCreator,
 } from 'slash-create'
-import MediaManager from '../classes/MediaManager'
-import { Client } from 'discord.js'
-import playMediaInVoiceChannel from '../utils/playMediaInVoiceChannel'
-
-export interface Options {
-  // Optional alternative command name, default to "play" (useful for command alias)
-  commandName?: string
-  // Optional alternative command description (useful for command alias)
-  commandDescription?: string
-}
+import MediaManager from '../classes/MediaManager.js'
+import playMediaInVoiceChannel from '../utils/playMediaInVoiceChannel.js'
+import EnhancedSlashCommand, { EnhancedSlashCommandOptions } from '../classes/EnhancedSlashCommand.js'
+import Media from '../classes/Media.js'
 
 export interface SlashCommandOptions {
   sound: string
 }
 
+export type PlayCommandOptions = Pick<EnhancedSlashCommandOptions, 'throttling' | 'throttleCache'>
+
 const REPLAY_BUTTON_ID = 'play_cmd_replay_btn'
 
-export default class PlayCommand extends SlashCommand {
-  constructor(creator, protected discord: Client, protected mediaManager: MediaManager, options?: Options) {
+export default class PlayCommand extends EnhancedSlashCommand {
+  constructor(
+    creator: SlashCreator,
+    protected discord: Client,
+    protected mediaManager: MediaManager,
+    options?: PlayCommandOptions,
+  ) {
     super(creator, {
-      name: options?.commandName ?? 'play',
-      description: options?.commandDescription ?? 'Play a sound in your current voice channel',
+      name: 'play',
+      description: 'Play a sound in your current voice channel',
+      throttling: options?.throttling,
+      throttleCache: options?.throttleCache,
       options: [
         {
           type: CommandOptionType.STRING,
@@ -40,12 +46,13 @@ export default class PlayCommand extends SlashCommand {
     })
   }
 
-  async run(ctx) {
+  async run(ctx: CommandContext) {
     await ctx.defer(true)
 
-    const guild = await this.discord.guilds.fetch(ctx.guildID)
-    const member = guild.members.resolve(ctx.member.id)
-    const voiceChannel = member.voice.channel
+    const memberId = ctx.member!.id
+    const guild = await this.discord.guilds.fetch(ctx.guildID!)
+    const member = guild.members.resolve(memberId)
+    const voiceChannel = member!.voice.channel
     const { sound } = ctx.options as SlashCommandOptions
 
     // Check if message author is in voice channel
@@ -72,12 +79,11 @@ export default class PlayCommand extends SlashCommand {
           },
         ],
       })
-        .then(() => ctx.registerComponent(REPLAY_BUTTON_ID, async (btnCtx) => {
-          playMediaInVoiceChannel(voiceChannel, media)
+        .then(() => ctx.registerComponent(
+          REPLAY_BUTTON_ID, (btnCtx) => this.replayButtonHandler(btnCtx, voiceChannel, media)),
+        )
 
-          return btnCtx.acknowledge()
-        }))
-
+      this.throttler?.registerUsage(memberId)
       playMediaInVoiceChannel(voiceChannel, media)
     } else {
       return ctx.send(`No media found for query **${ sound }** :upside_down:`, { ephemeral: true })
@@ -91,5 +97,22 @@ export default class PlayCommand extends SlashCommand {
     return this.mediaManager.getBySearch(sound)
       .slice(0, 25)
       .map(({ name, filename }): AutocompleteChoice => ({ name, value: filename }))
+  }
+
+  async replayButtonHandler(buttonCtx: ComponentContext, voiceChannel: VoiceChannel | StageChannel, media: Media) {
+    const memberId = buttonCtx.member!.id
+
+    if (this.throttler?.isThrottled(memberId)) {
+      this.onBlock(buttonCtx as any, 'throttling', {
+        throttle: this.throttleCache,
+        remaining: this.throttler!.getRemainingDuration(memberId)! / 1000,
+      })
+      return buttonCtx.acknowledge()
+    }
+
+    playMediaInVoiceChannel(voiceChannel, media)
+    this.throttler?.registerUsage(memberId)
+
+    return buttonCtx.acknowledge()
   }
 }