+++ src/types/Cache.ts
@@ -1,0 +1,6 @@
+import { ThrottleAllowedCommands } from './Config.js'
+import Throttler from '../classes/Throttler.js'
+
+export default interface Cache {
+  throttles: Record<ThrottleAllowedCommands, Throttler | undefined>
+}