+++ CHANGELOG.md
@@ -1,7 +1,32 @@
 # Changelog
 
-## [Unreleased](https://github.com/super-bunny/pld-cli/tree/dev)
+All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
-### API
+## 0.2.0 (2023-05-08)
 
-- Add `search` query parameter to `/sounds` endpoint to make search on sound list.
+
+### Features
+
+* **commands:** add status slash command ([3dba617](https://github.com/super-bunny/discord-soundbox-bot/commit/3dba617ff049c225bd77c76821aa429862b9d136))
+* print app env at startup ([af37fd5](https://github.com/super-bunny/discord-soundbox-bot/commit/af37fd539e7dff48886a1bce1a0e70a3e1bedbae))
+* **commands:** add media playing commands throttling ([1ba7f27](https://github.com/super-bunny/discord-soundbox-bot/commit/1ba7f271272e09f1dc4cd887a5ca78c1fa56b9b2))
+* **api:** add /stream route to get sound file stream ([851e471](https://github.com/super-bunny/discord-soundbox-bot/commit/851e471bc62af87b1dab2efb9723ad3f521b364e))
+* **api:** add auth cookie middleware ([ccc21d7](https://github.com/super-bunny/discord-soundbox-bot/commit/ccc21d78ed490b329905c76099e81ffcbf081fc8))
+
+
+### Bug Fixes
+
+* fix broken slash commands ([62eec13](https://github.com/super-bunny/discord-soundbox-bot/commit/62eec1304e480a301622bc62935ce266a1091d03))
+* fix broken bot by updating dependencies ([24cff7b](https://github.com/super-bunny/discord-soundbox-bot/commit/24cff7b7aa217da5b98cd631b41902fe8931a8a2))
+* **throttling:** add missing event register in play and random command ([f30c6b1](https://github.com/super-bunny/discord-soundbox-bot/commit/f30c6b1189a0fb62bdcd8e76143ad86e0a5c3915))
+* CONFIG_FILE env var type check ([333fcfe](https://github.com/super-bunny/discord-soundbox-bot/commit/333fcfe7c4a8f25ef07953ece1d2d68e990665f8))
+* **config:** write default config in file at startup ([425f3f9](https://github.com/super-bunny/discord-soundbox-bot/commit/425f3f908c0b58a49eebd7bb3135fb96f4ca5836))
+* delete commands on sync ([8a61725](https://github.com/super-bunny/discord-soundbox-bot/commit/8a617257cc99df3996cc261cd963efda9f64f636))
+* **api:** fix returned status codes ([958b996](https://github.com/super-bunny/discord-soundbox-bot/commit/958b9966809e41c586c71e0325b973e1aba6ae84))
+* @discordjs/voice encryption library not found ([4afb676](https://github.com/super-bunny/discord-soundbox-bot/commit/4afb676bc3e7f66415a9857e41579e651c8cc725))
+
+
+### Dependencies
+
+* update typescript package ([0992218](https://github.com/super-bunny/discord-soundbox-bot/commit/0992218ca041dcf5384f9b19766d9688dfdba3b5))
+* update body-parser package ([cda722a](https://github.com/super-bunny/discord-soundbox-bot/commit/cda722ac95aa2d6070b382811ef7ff465bc6e316))